refactor(useActivities): useActivities 파일 위치 변경
